refactor #154: move info and badges components to atoms
