+++ components/atoms/badges/__tests__/index.tsx
