+++ components/atoms/info/__tests__/index.tsx
