+++ components/atoms/info/index.tsx
