+++ components/atoms/info/styled.ts
