+++ components/columnCardsWithControls/index.tsx
@@ -3,7 +3,7 @@
 import { BestTime } from '@/interfaces/player'
 import ColumnCards from '../columnCards'
 import { PageContainer } from '../pageContainer'
-import Badges from '../badges'
+import Badges from '../atoms/badges'
 import { BadgesContainer } from './styled'
 import { useState, useEffect } from 'react'
 import { Difficulty } from '@/interfaces/difficulty'
